refactor, get prices by contract from coin gecko if possible, blacklist AAVE

+++ app/src/commonMain/kotlin/utils/math.kt
@@ -4,6 +4,6 @@ import com.ionspin.kotlin.bignum.decimal.BigDecimal
 import kotlin.math.pow
 
 
-fun normalizeDenom(amount: BigDecimal, decimals: Int): BigDecimal {
-    return amount.divide(BigDecimal.fromDouble(10.0.pow(decimals)))
+fun normalizeDenom(amount: BigDecimal, decimals: UInt): BigDecimal {
+    return amount.divide(BigDecimal.fromDouble(10.0.pow(decimals.toInt())))
 }
